Move reverse cursor into a module.
Move `ReverseCursor` into a separate module, so it can be reused more
easily.  Also, rename it to `ReverseKeyCursor`, as we may need a reverse
value cursor in the future.
