+++ crates/dbsp/src/trace/cursor/mod.rs
@@ -15,6 +15,7 @@ pub mod cursor_empty;
 pub mod cursor_group;
 pub mod cursor_list;
 pub mod cursor_pair;
+mod reverse;
 
 #[derive(Debug, PartialEq, Eq)]
 enum Direction {
@@ -26,6 +27,7 @@ pub use cursor_empty::CursorEmpty;
 pub use cursor_group::CursorGroup;
 pub use cursor_list::CursorList;
 pub use cursor_pair::CursorPair;
+pub use reverse::ReverseKeyCursor;
 
 /// A cursor for navigating ordered `(key, val, time, diff)` tuples.
 pub trait Cursor<K, V, T, R> {
